requirements/dev-venv.txt
bleach==5.0.1
boto3==1.24.51
botocore==1.27.51
build==0.8.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
commonmark==0.9.1
docutils==0.19
idna==3.3
importlib-metadata==4.12.0
jmespath==1.0.1
keyring==23.8.2
packaging==21.3
pep517==0.13.0
pkginfo==1.8.3
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
readme-renderer==36.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
twine==4.0.1
urllib3==1.26.11
webencodings==0.5.1
zipp==3.8.1
